adds knex-cleaner as a dependency for staging and production environments

+++ data/seeds/01-cleanup.js
@@ -1,4 +1,13 @@
+const cleaner = require('knex-cleaner');
 exports.seed = async knex => {
+  const environment = process.env.DB_ENV;
+  if (['staging', 'production'].find(environment)) {
+    return cleaner.clean(knex, {
+      mode: 'truncate',
+      restartIdentity: true,
+      ignoreTables: ['knex_migrations', 'knex_migrations_lock']
+    });
+  }
   await knex('tickets').truncate();
   await knex('categories').truncate();
   await knex('users').truncate();
